Up: move Action to class
