+++ app/turbulence/gcloud.rb
@@ -21,25 +21,6 @@ module Turbulence
 
     module_function
 
-    Action = Struct.new(:id, :name, :class_name) do
-      def to_choice
-        {
-          name: name,
-          value: self
-        }
-      end
-    end
-    def action
-      choices = Actions::LIST
-                .map { |action| Action.new(action::ID, action::NAME, action) }
-                .map(&:to_choice)
-
-      action = Menu.auto_select('Select your desired action:', choices, per_page: choices.length)
-      Config.set(:action, action.id)
-
-      action.class_name
-    end
-
     # rubocop:enable all
 
     def go!
